requirements.txt
django>=6.0.3
djangorestframework>=3.16
django-cors-headers>=4.7
